refactor: query로직 view로 이동
